refactor : seald class -> seald interface
